Use transform on router instead of margin-top on body
This change specifically sets what we need to have the spacer between
the navbar. This way we are not setting everything to have margin top
and then re-a-fixing the navbar to the top. Instead we are just grabbing
the content and moving it down the necessary amount. Also better for
mobile and performance.

AE-1944 #time 10m