Remove trailing slash on makeGet request
Since we added the unique pao session request we want to remove this
trailing slash which will allow us to use an empty string allowing us to
use the pao request successfully. Since the pao request is a get but you
do not specify a uuid. This allows keeps the other get requests
functioning more or less the same.

AE-2091 #time 15m