Use self signed cert for dev
AE-2132 #time 20m

Since this is required for development for now we can add it to the
repo.  They are just self signed certs that allow us to run in in
a development env as ssl.  Currently this is required for building on
Jenkins until we come up with a better way to handle this.