Add webpack define plugin
AE-2115 #time 20m

This allows us to define global variables to emulate process.env.  This
is requires so we know what config to load.  Since this solves the
version issue we dont have to regenerate the version.js file each time
and we can just pull it from the env variable.

This also adds the dedupe plugin which will remove duplicates when
webpack packages the requirements.