Change node env during jenkins build
AE-2115 #time 20m

We want jenkins to build the project in the correct environment.  If the
branch is on master and commits are zero for the current version we want
to build production.  If were on master without zero commits we want to
build development.  Otherwise we are just building a branch.