Add translations for stock form
AE-2157 #time 20m

Adds translations for stock form.  This has some issues as I am
translating the value of the dropdown options.  Since bootstrap only
allows strings or numbers for children of an option tag we are getting
a ton of warnings.  This should be fixed once we get an autocomplete
control instead of dropdowns.