Add container for adding/editing users on records
This handles sending a specific get request and/or sending a post/put for
updating users.

We then send out a post to the membership endpoint to
add the user we just created to a group setup in the Config.GROUP.

Also we simply remove the user from the group instead of deleting the
user completely on delete. Currently Permissions doesn't allow this to
work but will be changed in the future.

AE-2169 #time 4h