Display prints on edit order page
The print list has just two column, id and status. When we add prints to
rapidfab, we will be able to click on the id of each print and see its
details. I haven't fetched the runs associated with a print because we
don't have a direct way to do that in wyatt.

AE-1955 #time 1h