Add container for managing data and requests for New Order
This sets up the data we need for the New Order page.
This also handles sending the upload request and trackings its progress.
We also setup the payload for uploading as well as when its finished we
redirect to the newly created page with the orders information filled
it.

AE-2174 #time 2h