Fix api middleware error handling
AE-2237 #time 1h

Since the /run/ list endpoint did not exist it was returning an HTML
payload of 405.  We need to be able to handle whether we can parse the
response of the payload into json or not.  Regardless if we can or not
we need to dispatch an event based on error or success.