Add edit run page
The page displays the estimates and actuals, status and created time of
the run. Only status can be changed. Incase there is invalid
transformation of the status, it will be displayed when the user
attempts to save the status. A list with prints and associated orders is
displayed in a table.

AE-2236 #time 1d