Add test to make sure LIST doesnt clear resource state
AE-2292 #time 20m

Originally AE-2292 thought that the reducer was clearing the state and
adding all the records back in.  This is not the case and it is behaving
how it should by only adding records to the state and overriding ones it
already has.

This commit enhances the tests to make sure we are testing this use
case.