Add count to ui method state
AE-2292 #time 10m

This allows us to track how many times an endpoint has been called or
tied to call.  The simple reason to use this is to make sure we don't
have to list the endpoint each time the action is called.  If the list
has already been ran once we dont have to do it again.  If we have
"caching" issues later we could make this a bit smarter using the event
bus as well.