Fix model upload error and progress bar
payload.model had the file value from the field when grabbing the file
they want to upload. We were then overwriting it with the location
header for the model before uploading it. Now we upload the model then
assign the model uri into the payload for our order request. Fixing both
our progress bar and the model upload.

AE-2286 #time 2h