Fix erroring action calls and fix the upload action
Currently when running upload Progress it is constantly being ran
whenever the upload progress is updated. So it doesn't finish with
request args like I first imagined it would. So we are simply taking
these out for now in the future we may have further implementation of
tracking error handling or when it is complete with an action. Currently
we do check the prior request and the one there after and then also we
have a statement in the container that checks when the progress is
finished and acts accordingly.

AE-2286 #time 1h