Remove undefined actions in our log
The reason we were getting these actions is because the API middleware
translates from a rough prototype for some actions that should be
dispatched into two or more actual actions. That prototype itself is not
an action and the middleware eats it. However, since the logger
middleware was initialized in the chain first it was capturing the
action prototype and logging it. Now that logging happens last we only
actually log things that go to the stores which is much better

AE-2304 #time 30m