Stop attempting to render a text box with null value
So the issue here is that the null value raises an error in
bootstrap-react. We are using redux-form which is getting its data
directly from the API. Rather than doing some data munging of the data
coming back from the API and going into the store I decided the real
issue was sanitizing our values before display. I created a new widget,
FormControlTextCareful which accomplishes this by translating from null
to an empty string for display purposes. This gets rid of the warning
from react-bootstrap

AE-2304 # time 1h