Filter out bad props for text areas
We were getting a warning about the many props that redux-form is
passing through to BS.FormControl that it hasn't whitelisted. Rather
than filter those values out we're going to just change the display
element and use it to filter them out. This gets rid of the warning

AE-2304 #time 30m