Show the run ID on the run page
In order to do this I had to bypass the default behavior for all
resources coming out of APIs - to overwrite the 'id' property with the
last 6 characters of the UUID. I need the UUID to be unique, so we need
to show a full UUID so the user will type in a full UUID and get
reliable linking between the run and machine analytics.

This also means in lots of places we'll display really big run IDs.
That's ok for now.

AE-2314 #time 20m