Show loading indicator rather than force full page wait
On the order page we were making the entire UI wait for a listing of all
of the models. It's a much better user experience to show the data we do
have as we have it.

Now we do that.

If the model data isn't loaded we show a loading indicator instead of
the select box for the model. As soon as we have all the model data we
show the list

AE-2324 30m