Translate chart.js labels
AE-2377 #time 2h

This allows us to use react-intl to translate the labels for the charts.
Since chart.js uses a canvas we cannot just render the labels as react
components so we use the injectIntl function of react-intl.

This also cleans up the status keys for i18n in the transation files.