Fixup uploadModel reducer
AE-2454

We need to know what model is being uploaded after it has been created.
This way we can watch the status from the resources part of the state to
base the UI off of what the status is of that model.

This also includes tests for the reducer and its new changes.