Add clear state and save order payload actions
AE-2454

We need a way to store the current forms payload while we wait for the
model to process.  Then we can create the order.  This stores the
payload for use later.

We also need a way to clear the uploadModel state so we can start it
over again after an upload completes.