Ignore actions that have no payload
This happens because on connect to eventbus we send down an event
indicating that the client has connected (since otherwise they'd have no
real good way to know that things are fine). This event doesn't include
a payload on purpose because it's a meta-event. We just need to throw
them away. Hopefully this won't burn someone in the future who wants
these events :)

AE-2486 #time 20m