Import correct i18n lib
AE-2482

We are overloading react-intl to provide custom formatted components.
In some cases we were importing FormattedDateTime from react-intl but it
should be from rapidfab/i18n since react-intl doesnt include
a FormattedDateTime component.

This also makes the component allow a string along with an object as the
value prop.