Add code to perform the model get and parse it
I had to create a new function for doing the request. I put this in
makeApi so that it could follow similar patterns to what is already
there and established and so that in the future we wouldn't lose track
of it. I also exported jsonTryParse so that I'm not reproducing the JSON
parsing code from the API middleware. I would have exported something
higher level but everything else is deeply tied into the middleware
types

AE-2315 #time 1h