Properly use API machinery to do model download
This is the result of a long conversation with Nick about how to
properly leverage the API middleware to make this logic happen. Rather
than directly using doGet and jsonTryParse we instead emit an action
which actually tells the middleware how to do the request. The
middleware still handles the parsing and we just use the resulting
promise to download the model.

I thought we'd have to do something elaborate to get the browser to
download the file, but turns out that after talking to Nick it's
rediculously easy. Just set the window location to the content of the
file and the browser will do the right thing. Technology, right?

AE-2315 #time 3h