Show user feedback about the model download
I decided to revert my previous changes and go back to having a separate
action for downloading the model. The major reason is so that I can show
UI information to the user and give them some feedback that we are
working on the download they requested. This has the nice side effect of
avoiding multiple successive attempts. This is important as our users
get farther away from our API and there is noticeable lag when getting
the model content link.

This also maintains the hard divide between display-only UI and actions
which I seem to have convinced Nick of

AE-2315 #time 1h