Remove ability to modify from/to specific statuses
We need to limit what the user can change a status to and what they can
change a status from. Currently queued, calculating and calculated are
a status that the service itself updates a run to. From there our
service will handle updating that to whatever status follows.

A user should not be able to switch to these statuses manually or switch
from them and doing so could mess up the flow of the current run
causing problems.

AE-3101 #time 3h