Query for all runs
AE-3133

We were only querying for runs that were in a queue for a machine.
Since a completed run is not in a queue we were not querying for them to
render.  This queries all runs.  In the future we will want to add in
date range filters so it does not get out of hand.