Update logic for deleting a user
This just removes the user from the bureau's group. Previously we were
using the .get() function which, as far as I can tell does not support
filtering. The results coming back also do not have a direct uri
property so we pull through the results coming back and find the URI so
we can get the UUID we need

AE-3208 #time 20m