Add the ability to manually remove an object from the API store
Okay, this was a doozy. The problem here is that after we delete the
user we didn't delete them from the resource store. This is what
controlled the table of users. So we had to plumb a way to tell the api
cache that we need to evict a record. This commit does that.

AE-3208 #time 1h