Show TOS component if user has not agreed
AE-2359

Show TOS component if the current user has not agreed to the terms.  If
they have already agreed then we don't care to require them to agree.
We also dont want to show the component if the current user is being
impersonated.  This way we can still use the app even if the user has
not agreed to the terms yet.  Which is nice because we don't want any of
us agreeing for someone else.