Add new action for grabbing data from order-location endpoint
We need a custom action to grab the data from order-location. We cannot
use the api meta-programmed actions because this endpoint doesn't return
a uri and that set of actions uses the uri as a key when adding it to
the state. So we needed to create our own, grab the data and save to the
state.

AE-2527 #time 8h