Avoid emitting a warning about our missing translation
The difficulty here is that we have a formatted message that we are
telling the translation system has an ID of uxNavBrand, but we never
really intend to use the static translations for it. We instead are
abusing the defaultMessage to set it dynamically. This will always give
us a warning. Instead we're going to just dynamically set the name to
avoid the warning

AE-3582 #time 5m