Request API calls properly inside the onInitialize member
without this we emit an error related to setting the state during an
existing state transition. This is because we are requesting functions
that are dispatching actions which are in turn modifying the store. We
shouldn't be doing that during a render loop. By tucking this away in
the onInitialize we can avoid the warning entirely

AE-3582 #time 5h