Use the order payload to create a simple flag for creating orders
In production we're hitting a problem where we are creating an order
ever time we go through the render loop. This is because react attempts
a re-render 10 times before giving up to see if the state will settle
out. This creates 10 copies of our order.

Now instead, we reset the order payload to null as soon as we attempt
the creation so that we won't re-enter the same code and create order
duplicates. We'll want to find the cause of the original re-rendering
loop, but until we get there this will work. This will also get reworked
whenever we do the order upload in the background

AE-3943 #time 1h