Refactors Order Form by adding template input fields
The template input fields replace the post-processor fields. The changes are
namespaced in the temp directories and routes in order to hide the new features
in production. In addition, it was noted that the new order form does not use
redux form whereas the edit order form does; see AE-3955.

AE-3937 #time 8h