Adds UI form for adding and editing process steps
On the page for adding and editing process templates, users can click to add
a process step or edit a process step. Clicking the buttons will bring up
a modal with a selection input for process type and radio buttons for
additional data options. This commit adds UI that is not all fully functional.

AE-3947 #time 4h