Connect duplicate UI to API
The onDuplicate function in the Template container takes a templateCopy object
from the Template component. In order to duplicate the template, the function
creates an array of promises resolving post requests for each copied step. Once
the step API requests have returned, the function posts a new template with the
copied steps.

AE-3950 #time #6h