wire up all the templates!
the work ended up being way too big for one ticket, but here we are. This ticket involves wiring up process steps and templates to the api.
It also includes a ton of bug fixes as I was wiring things up and found stuff.

AE-3936 #time "what year is it?"