Unroll python logic into Jenkins workflow
This is pretty complex but I think it will work better than using the
python 'jenkins' script. Instead we'll directly use the logic we have at
the Jenkins layer to conditionally deploy. This involves less risky
communication with git which doesn't actually have a good way to move
out from a commit to the branch it belongs to reliably