Add print container and component
The print component is static and intended as a prototype of the mockup, not
a final product. Adding dynamic data and connecting to the API will be
addressed in a future commit.

The print container and component are designed and styled slightly differently
than the current standard container and components.

The print container now handles all error and loading rendering, rather than
the component. This should improve the separatation of UI concerns and enhance
reusability.

The print component puts a .container class on the grid layout, which adds
responsive sizing. The back button on record components has been removed
because it was redundant with the breadcrumb navbar right above it, which can
cause confusion in users.

AE-4040 #time 8h