Fix order form status field bug
The status field was rendering the options based upon the value in the redux
store. This caused an unintended side effect where if the user selected
a different option, changing the store, the options in the select field were
different, and the user could not undo the change.

Now, the order form is a stateful component that remembers the initial
statusOptions in state for its lifecycle.

AE-4042 #time 4h