Fix run update function
Redux-form was sending all run data back to the API. The API would throw an
error because it does not allow a run status be set to the same value. Now, the
onSubmit function checks the initial state of the run status. If the run status
has not changed, it does not send the run status to the API, which prevents the
error.

AE-4081 #time 8h