Decouple edit order form logic from line item form logic
In order to manage the order form and the line item forms, we decouple the form
logic for both. Now, the EditOrderForm is responsible for its own state, and
the LineItem is reponsible for its own state, which allows for composability
instead of inheritance.

The new features are under hidden routes until deployment or feature flags.

AE-4159 #time 8h