Create logic for line item form and refactor
This commit refactors the LineItem component by creating a new LineItemForm
component to handle all of the form logic, and all of the other view displays
handled by the LineItem component.

The onSubmit and onDelete functions are not implmenented as of yet, pending the
Line Item API to be completed.

This commit also includes various refactors for reducing unecessary code and
imports.

AE-4164 #time 6h