Fix loading bug and refactor Printable component
In order to fix the initial render bug, when determining if a model is
printable, the conditional statement does not try to dig into data that might
not be there. This is done by adding the addditional conditional '&&
model.analyses.'

In addition this commit refactors the conditional that handles the rendering by
adding better UX for ITAR models (displays 'unknown' instead of 'loading') and
by refactoring out the nested conditional statement for readability.

AE-4275 #time 1h