Run prettier on js files
Prettier minifies js to remove formatting and then reformats it for us.
This allows us to not worry about formatting code style, since prettier
will do it for us.

In addition, this fixes all of our formatting eslint errors, instead of
having to do it manually, which is boring.

AE-4616 #time 1h