Remove null sales_status payload on order put
Sales status is a unique case. It is under a feature flag, which
sometimes prevents a default value being set, and cannot be null. The
order does know whether the bureau has the eos-order-fields feature flag
enabled, so it has to assume that sales_status is a potential value, and
will default to null if nothing is selected.

The API will reject a put request with sales_status having a null value.
So, get rid of it!

AE-4677 #time 30m