Remove EditOrderForm container component
The container was intended to hold the status options based upon the
initial value of the status. This, however, is not good style!

The status options should not be held in state, but rather just
calculated on the fly with intialValues.

This allows us to remove the entire container! Yay, remove lots of code.

AE-4677 #time 30m