Rework docker image generation
We figured out in our testing that because we set up the rapidfab
container once and then start and stop it that we weren't actually using
the latest code when building on feature branches. That's wrong. This is
an attempt to change that by only building the image if we don't have
one locally, which is not much different from what we were doing, but
always starting and stopping and *removing* the container based on the
image. That's the new hotness

AE-4606 #time 30m