Install npm in normal mode, copy modules to a safe place
This is kind of hackish, but I think it'll work. When I install
globally, no matter how I fiddle with the options, I don't seem to get
all the dev dependencies like karma. So instead I'm going to install
normally and then just save off the data to another location. This will
increase the size of the image, but I don't care at this point. I then
tell jenkins to link up the saved node modules to the source directory
and it all should work.

Should.

AE-4606 #time 20m