Run eslint --fix
My local eslint config was not in sync with the local eslint config.
Now, instead of running prettier through a separate module, it is
integrated with our eslint config, so we can pass the --fix argument to
the eslint command and it will fix all of the prettier errors. Wow!

AE-4606 #time 10m