Explicitly specify the files to put in the image
We need this because right now we are adding all of the current working
directory, including node_modules. That's wrong on two counts. First,
it's wrong because we don't want to push that much data through to the
image it makes it big and accomplishes nothing. Second, on successive
builds that means we'll actually nest the node_modules inside
node_modules ad infinitum