Add download button for traceability report
We are using the a href download element instead of the javascript
window.open function, which is frequently blocked as a popup.

In addition, we add some loading state to the TraceabilityReport
component so that we can have a loading state for the button while the
report is being generated.

AE-4718 #time 1h