Remove the invisible bureau input field
The field had a style 'display: none' which rendered it invisible. I'm
not certain what the purpose of this was, but it is not needed anymore
at the very least.

Mapping over the bureaus was causing the route to break. So just get rid
of the code that we aren't using.

AE-4740 #time 10m