Show bureau error based on roles, not presence of bureaus
This is the major work of this feature branch. We want to detect whether
or not a user has a role in a bureau, not whether or not they can see a
bureau at all, so that we can control what features they use based on
their role. This means we need to store the role(s) in the session for
later retrieval as well as erroring based on the roles they have on
different bureaus.

AE-4846 #time 2h