Update API to return promise when suppressing request
The issue is this - now that we supporess additional requests when a
request has already been made we need to make that behavior the same
as when we do a real request. That means returning a promise so that
calling code can know when the request has (not) been made and can
respond accordingly. This also means we need to store the previous
result and return it as part of resolving the promise.

I also took a moment to refactor the middleware so that the events being
dispatched are defined in a single place that we reuse. This ensures
they all have the same basic properties plus their own additional
properties depending on the type.

AE-4464 #time 2h