Fix Edit Order loading bug
The origin of this problem is in two different patterns for passing the
order data down to the Order components. Previously, the page container
would pass down the data as props down to all components.

I am moving us more toward connecting the order data to components
through the redux store, not through props. This benefits us by
decoupling components, and simplyfing our component dependencies.

Also, this is what redux is for, so let's do it.

Unforunately, passing down the route resource through redux is not fully
implemented yet.

As of now, this solution is hopefully only temporary and will get refactored out once we have the route resource available through a selector.

AE-4894 #time 1h