Don't close the edit/add user modal on random events
Whenever we get an event from the eventbus we are redrawing the page. If
we don't have new state or new props all this does is close the modal,
which is super annoying. Instead we now explicitly check and if we don't
have a reason to re-render, we don't

There's probably a more efficient fix for this

AE-4464 #time 20m