Disable and blank location selector for non-location roles
It was just confusing to allow a user to change the location on global
roles. So now we blank it out. We also select the first location when
the user changes roles so that we have an actual populated value if the
user just selects "location user" role and doesn't change the default
selected location. This is safe because in the logic for sending the
payload we screen out the location for all other roles types

AE-4464 #time 20m