Update onCreateUser to POST /role if user already exists
If the POST /users fails because the user already exists, we still need
to POST /role.

Catching the error is not the most elegant way to do this. Am I proud
of it? Not really. Does this need to go out? Yeaaaaaaaaaah

Ideally, we will be able to refactor this Dashboard code and remove this
soon.

AE-4464 #time 30m