Remove unnecessary LIST /print request
And this solves the problem. The actual LIST /print request that we want
to send is below, and was getting suppressed because we were sending
duplicate requests.

The LIST /print request below contains the filter on line_item that we
need to use, so that's the actual important one.

This PRINT /list request was redundant and suppressing the actual one.

AE-5013 #time 1h