Create .eslintignore and add i18n data to it
cause the i18n data needs to be JSON like, which the linter doesn't want

also get rid of the lint:js-output script in package.json cause we don't
need that anymore. Was in back when we had >1000 eslint errors and need
to search through them easier. Don't need it anymore =D

AE-5118 #time 5m