Create getProcessStepsForPrints selector
In order to get the name of a print step, we need to look at the process
step, then at the processor_type, which means we need to traverse and
iterate over multiple arrays of resources.

Sounds like a good job for a selector so that the result will be cached
and we don't need to recalculate on every render.

AE-5314 #time 1h