Refactor reducers/api to use extractUuid util
This was found while investigating the absence of order documents.

The order documents were not displaying because the LIST/ features
response data was not entering the redux store.

The data was not not entering the redux store because one of the
features had a null uri value, which the reducer tried to parse unsafely
and broke the entire function.

The reason that the reducer was trying to parse the value unsafely was
because this extractUuid function is duplicated in our codebase, and
this version of it did not properly verify the parseability of the uri.

So, if we remove this duplicate function and replace with the function
in the utils folder, which has proper null checking, we don't break the
reducer function and properly get the response data in our store.

AE-5326 #time 2h