Fix mapStateToProps bug in PrinterFormContainer
This was found while investigating errors when trying to create a new
printer.

The printer form was submitting null values to the API, instead of
default values.

It was sending these null values because sometimes, depending on the
properties of the route prop, the getRouteResource selector would
return an empty object instead of null when the route did not have a uuid in it.

We later check whether the printer value is false in order to decide
whether the PrinterForm is being used as a New Form or an Edit Form
for an existing printer.

In JS, an empty object is not falsy, so this empty object was telling
the PrinterForm that is was editing a printer and thus was expecting to
receive values that it never got. So, we need the printer value to be
falsy when the props.route.uuid is null or undefined.

I thought the selector did this, but I was mistaken.

So, instead, explicity check for this condition.

AE-5327 #time 1h