Fix User email display bug
Found while investigating sentry error

The sentry error said there was an object as a child of a react component, which is not allowed.

This was the only potential object being rendered on the component, so
hypothetically the emails key could have had an object not an array
inside of it, which would have caused such an error.

So, in order to prevent that from happening, we should not just check forr
 the existence of the user key, we should check to make sure that is an array
before indexing into it.

AE-5339 #time 1h