Force chunked requests to avoid debouncing/suppression
Without this we only honor the first request of the chunk. This gives us
a random sampling of the line-items, models and prints that we should be
able to use for a new run.

That's bad.

The real solution would be to add respect for filters to our debouncing
logic. But, that implies our debouncing logic is good and should be
maintained. It is not. Long-term we shouldn't have the debouncing logic,
we should respect caching headers at the browser level. But that would
require fixing our render loop

:(

AE-5357 #time 3h