Fix new manufacturer form bug
The redux-form did not have a bureau value initialized when creating a
new manufacturer. The API requires a bureau value to be supplied when
POSTing a new manufacturer.

So, to fix the problem, we initialize the bureau value in redux-form
explicitly.

I'm not 100% certain what caused this bug in the first place, this code
hasn't been touched in a while. My guess would be an API change or side
effect from some other code change.

AE-5388 #time 30m