Show events from runs and prints
First I need to translate the reference type so that we get a nice
display of the properties.

Second I need to allow events for properties that went from None to
something because all of the important run events do that.

This makes the page huge.

Oh well.

AE-5389 #time 15m