Fix handleError bug in middleware/api
Turns out that the errors argument can be either an array or an object

I was assuming that the argument was always an object, and only handling
that path, which caused problems because it coerced an array into an
object which didn't work.

So, check if the errors argument is an object before trying to
manipulate it.

AE-5413 #time 30m