Remove componentWillReceiveProps from Run new
This function was resetting the selected printer whenever the component
received props from the event stream, so I removed it.

I'm not 100% certain what the original intention of this lifecycle
function is after testing with it removed.

The function checks to make sure that the selected printer is in the
printer props array, but I'm not sure why we need to this.

If for some reason this does occur, the submission will throw an error
and alert the user, so having this check seems unnecessary.

If for some reason this function is necessary for a reason I don't fully
understand as of right now, I am fairly certain that there is
alternative implementation that has a clearer intention and purpose.

But right now, I'll just assume this is bloat code that can be safely
excised.

AE-5546 #time 30m