Always build rapidfab image, always tag it
This is essentially undoing my previous optimizations to only build the
image occasionally in favor of building always so that we bake in all of
the assets.

I'm also avoiding overwriting the node_modules directory by directly
mapping all the source code and instead mapping only a couple of
subdirectories.

This is probably the wrong way to do this frankly since all the code is
baked in.