Exclude config file from the bundle
This will allow us to override it specifically in different
environments. The content remains the same and it should be available
through nginx (though we don't currently upload it to S3). I have
uploaded it manually so that the web pages still work (hopefully).

Eventually we're not going to use CloudFront any more so it won't really
matter