Update references to COMMIT_HASH and BUILD_VERSION
These weren't properly being set because we are trying to access them
during the docker image build (which we weren't doing a week ago).
Docker doesn't have these environment variables set the way that node
did when it used to build so a bunch of things broke around our
commit-specific actions. Now I've updated the jenkins environment so
that it properly sets the variables, and renamed the BUILD_VERSION to
GITDESCRIBE as in other projects for clarity

I'm not sure if this works until it goes to Jenkins, so here's hoping