Add detailed error information on 'failed to fetch'
This serves multiple purposes. The primary purpose is to change the
message pattern so that all of the generic errors in sentry don't get
bucketed into the same location so that we can't cleanly resolve and
deal with each individual case.

Also this will help our users to give us more specific information.
'Failed to fetch' tells us very little. This will at least tell us now
which API and what the status code was

AE-5686 #time 30m