Fix promise chain bug in print record
We fix the promise chain by conditionally sending requests that could
possibly fail under certain conditions.

The LIST /event was never being requested because the promise chain was breaking.

The promise chain was breaking because the request for run documents was
returning an error, breaking the chain.

The run-document request was failing because sometimes it was sending a
filter with an empty value.

This was difficult to see, however, because this promise chain is fairly
lengthy and convoluted.

Now, we should be covering all of the possibitilies.

AE-5732 #time 1h