Fix AddLineItem baseMaterial state default
Sometimes, when a user goes to add a line-item, the default base
material was not being set correctly, causing the POST/ line-item
request to fail because it is a required value

The line-item default was not being set because of a race-condition
between the GET/ order and the GET/ material. If the GET /order returned
first, or the order information was cached, the AddLineItem component
would construct, but not have a baseMaterial to set as a default.

We fix this by adding the componentWillReceiveProps lifecyle component
that can set the default baseMaterial state once the GET /material
request succeeds, instead of just relying on the constructor.

AE-5991 #time 30m