Add material informatino to material stock form
When a user selects a material for a material stock, they should be able
to see information about the material for a better experience.

Specfically, the user should be able to see the unit measurement of the
material so that when they enter a value for the quantity they know what
unit of measure to use.

AE-6039 #time 2h