Add none option to business_segment in order
Because coming from the API, some orders don't have a business segment
value. Previously, we would deal with this by just having the default
display segment be the first segment.

Now that users are downloading data directly ffrom the api, with the CSV
report, these null values are displayed in the csv, even though they
display differently in the UI

So, add a 'none' option because that is what we are getting from the
API, so this will keep our UI and API more in sync.

AE-6114 #time 1h